# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Whitfield WMS API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/v1/users": {
   "post": {
    "tags": [
     "Users"
    ],
    "summary": "Create User",
    "description": "Create a warehouse staff user (admin only).\n\nArgs:\n    request (UserCreate): User creation payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    UserResponse: Created user.\n\nRaises:\n    HTTPException 400/403/500.",
    "operationId": "create_user_v1_users_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Users"
    ],
    "summary": "List Users",
    "description": "List users.\n\nArgs:\n    role (Optional[str]): Role filter.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[UserResponse]: List of users.",
    "operationId": "list_users_v1_users_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "role",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Role"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/UserResponse"
         },
         "title": "Response List Users V1 Users Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/users/{user_id}": {
   "patch": {
    "tags": [
     "Users"
    ],
    "summary": "Update User",
    "description": "Update a user.\n\nArgs:\n    user_id (str): Target user id.\n    request (UserUpdate): Update payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    UserResponse: Updated user.",
    "operationId": "update_user_v1_users__user_id__patch",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Users"
    ],
    "summary": "Delete User",
    "description": "Delete a user (admin only).\n\nArgs:\n    user_id (str): Target user id.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    dict: Confirmation message.",
    "operationId": "delete_user_v1_users__user_id__delete",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/v1/warehouses": {
   "get": {
    "tags": [
     "Warehouses"
    ],
    "summary": "List Warehouses",
    "description": "List warehouses.\n\nArgs:\n    auth (dict): Authenticated user claims.\n\nReturns:\n    list[WarehouseResponse]: Warehouses.",
    "operationId": "list_warehouses_v1_warehouses_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/WarehouseResponse"
         },
         "type": "array",
         "title": "Response List Warehouses V1 Warehouses Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "Warehouses"
    ],
    "summary": "Create Warehouse",
    "description": "Create a warehouse (admin only).\n\nArgs:\n    request (WarehouseCreate): Warehouse payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    WarehouseResponse: Created warehouse.",
    "operationId": "create_warehouse_v1_warehouses_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WarehouseCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WarehouseResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/v1/warehouses/{warehouse_id}": {
   "get": {
    "tags": [
     "Warehouses"
    ],
    "summary": "Get Warehouse",
    "description": "Fetch a warehouse by id.\n\nArgs:\n    warehouse_id (str): Warehouse id.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    WarehouseResponse: Warehouse.",
    "operationId": "get_warehouse_v1_warehouses__warehouse_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "warehouse_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Warehouse Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WarehouseResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Warehouses"
    ],
    "summary": "Update Warehouse",
    "description": "Update a warehouse (admin only).\n\nArgs:\n    warehouse_id (str): Warehouse id.\n    request (WarehouseUpdate): Update payload.\n    auth (dict): Authenticated user claims.\n\nReturns:\n    WarehouseResponse: Updated warehouse.",
    "operationId": "update_warehouse_v1_warehouses__warehouse_id__patch",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "warehouse_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Warehouse Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WarehouseUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WarehouseResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CarrierScheduleCreate": {
    "properties": {
     "carrier": {
      "type": "string",
      "title": "Carrier"
     },
     "pickup_time": {
      "type": "string",
      "title": "Pickup Time"
     },
     "days": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Days"
     }
    },
    "type": "object",
    "required": [
     "carrier",
     "pickup_time"
    ],
    "title": "CarrierScheduleCreate",
    "description": "Carrier pickup schedule."
   },
   "CarrierScheduleResponse": {
    "properties": {
     "carrier": {
      "type": "string",
      "title": "Carrier"
     },
     "pickup_time": {
      "type": "string",
      "title": "Pickup Time"
     },
     "days": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Days"
     }
    },
    "type": "object",
    "required": [
     "carrier",
     "pickup_time",
     "days"
    ],
    "title": "CarrierScheduleResponse",
    "description": "Carrier pickup schedule."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "OperatingHoursCreate": {
    "properties": {
     "open": {
      "type": "string",
      "title": "Open"
     },
     "close": {
      "type": "string",
      "title": "Close"
     }
    },
    "type": "object",
    "required": [
     "open",
     "close"
    ],
    "title": "OperatingHoursCreate",
    "description": "Warehouse operating hours."
   },
   "OperatingHoursResponse": {
    "properties": {
     "open": {
      "type": "string",
      "title": "Open"
     },
     "close": {
      "type": "string",
      "title": "Close"
     }
    },
    "type": "object",
    "required": [
     "open",
     "close"
    ],
    "title": "OperatingHoursResponse",
    "description": "Warehouse operating hours."
   },
...
   "UserCreate": {
    "properties": {
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "minLength": 6,
      "title": "Password"
     },
     "role": {
      "$ref": "#/components/schemas/UserRole"
     },
     "warehouse_id": {
      "type": "string",
      "title": "Warehouse Id"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "full_name",
     "email",
     "password",
     "role",
     "warehouse_id"
    ],
    "title": "UserCreate",
    "description": "Payload to create a warehouse staff user."
   },
   "UserResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "email": {
      "type": "string",
      "title": "Email"
     },
     "role": {
      "$ref": "#/components/schemas/UserRole"
     },
     "warehouse_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Warehouse Id"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "last_login": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Login"
     }
    },
    "type": "object",
    "required": [
     "id",
     "full_name",
     "email",
     "role",
     "is_active"
    ],
    "title": "UserResponse",
    "description": "User representation without the password hash."
   },
   "UserRole": {
    "type": "string",
    "enum": [
     "admin",
     "manager",
     "receiver",
     "picker",
     "readonly"
    ],
    "title": "UserRole",
    "description": "Roles allowed for warehouse staff users."
   },
   "UserUpdate": {
    "properties": {
     "full_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Full Name"
     },
     "role": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/UserRole"
       },
       {
        "type": "null"
       }
      ]
     },
     "warehouse_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Warehouse Id"
     },
     "is_active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Active"
     }
    },
    "type": "object",
    "title": "UserUpdate",
    "description": "Payload to update a user's non-credential fields."
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "WarehouseCreate": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "city": {
      "type": "string",
      "title": "City"
     },
     "state": {
      "type": "string",
      "title": "State"
     },
     "address": {
      "type": "string",
      "title": "Address"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active",
      "default": true
     },
     "carrier_schedules": {
      "items": {
       "$ref": "#/components/schemas/CarrierScheduleCreate"
      },
      "type": "array",
      "title": "Carrier Schedules"
     },
     "operating_hours": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/OperatingHoursCreate"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "name",
     "city",
     "state",
     "address"
    ],
    "title": "WarehouseCreate",
    "description": "Payload to create a warehouse."
   },
...
   "WarehouseResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "city": {
      "type": "string",
      "title": "City"
     },
     "state": {
      "type": "string",
      "title": "State"
     },
     "address": {
      "type": "string",
      "title": "Address"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "carrier_schedules": {
      "items": {
       "$ref": "#/components/schemas/CarrierScheduleResponse"
      },
      "type": "array",
      "title": "Carrier Schedules"
     },
     "operating_hours": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/OperatingHoursResponse"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "city",
     "state",
     "address",
     "is_active"
    ],
    "title": "WarehouseResponse",
    "description": "Warehouse representation."
   },
   "WarehouseUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "State"
     },
     "address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "is_active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Active"
     },
     "carrier_schedules": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/CarrierScheduleCreate"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Carrier Schedules"
     },
     "operating_hours": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/OperatingHoursCreate"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "title": "WarehouseUpdate",
    "description": "Payload to update a warehouse."
   }
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/v1/auth/login"
     }
    }
   }
  }
 }
}